+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDQW6KQTZ79D6KS95E34B87"
+        "01KFDS3C5812XXS4REXFC920S9"
     )
     r"""The unique identifier of the data source"""