+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDQW6KSCQPGDN54DM6JRNRN"
+        "01KFDS3C5D02EHZCXMRZY2J56T"
     )
     r"""The unique identifier of the data source"""