+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDQW6KX5Z03DK0BXPHG6VC2"
+        "01KFDS3C5T31C42K20AKF16S9S"
     )
     r"""The unique identifier of the data source"""